Rename tests and mocks to identify group
